+++ Touhou Spell Up!!!/Assets/Scripts/DanmakuGame/Utility/RotationDirection.cs
@@ -1,0 +1,12 @@
+using UnityEngine;
+
+namespace TouhouSpellUp.Danmaku
+{
+    public enum RotationDirection
+    {
+        [Tooltip("反時計回り")]
+        CounterClockwise,
+        [Tooltip("時計回り")]
+        Clockwise
+    }
+}